Update tsconfig, modify README and adjust test description
This commit includes updates to tsconfig.json to include different file types, a semantic change in the README.md's project description and a modification to the entity description in a test file. The inclusion of new files types in tsconfig.json will help with codebase scanning, and the changed entity descriptions will better reflect our system's design and purpose.