Create "altas_users" table with unique email constraint
This commit introduces several modifications including creating the "altas_users" table and adding an "email" column with a unique constraint. The changes are encapsulated in SQL migration files along with respective metadata. Each step in the process is recorded in the database migration journal. The table "users" is renamed into "altas_users" as part of the modification.