Relocate hashing service features to application level
The Argon2 hashing service and its corresponding unit tests have been moved from the "services" directory to the "application/services" directory. Moreover, a log statement is added in the constructor of Argon2HashingService to reveal the hashing algorithm in use. The paths referenced in these files were accordingly updated. This change helps to better organize our codebase and increase traceability of logged events.